DUNEatLapp

Documentation for DUNEatLapp.

DUNEatLapp.apply_E_resolution

DUNEatLapp.apply_std_E_resolution

DUNEatLapp.get_evts_index

DUNEatLapp.get_n_primaries

DUNEatLapp.get_rate_neutron_captures_Ar

DUNEatLapp.greet_DUNEatLapp

DUNEatLapp.apply_E_resolution — Method

function applyEresolution(TrueEdata::Vector, Eresolution::Int) Function to apply an energy resolution on data. This resolution comes from the MicroBoone experiment and it is by default at 10% for 1 MeV. To adjust the resolution, you have to take in account the 10 % and add or subtract to obtain the the desired value. It accepts a vector for the TrueEdata and an integer for the Eresolution. It returns a vector.

DUNEatLapp.apply_std_E_resolution — Method

function applystdEresolution(TrueEdata::Vector,Eresolution::Real) function to apply a sigma/E = resolution / sqrt(E) energy resolution It assumes that the TrueEdata is in MeV, otherwise the results will be incorrect

DUNEatLapp.get_evts_index — Method

function getevtsindex(df::DataFrame) Function to get the index fo start and end of each event in a ulalap.csv file It accepts a dataframe and returns a matrix with column corresponding to the: numberofevent indexstartevt indexendsevt

DUNEatLapp.get_n_primaries — Method

function getnprimaries(my_file::String) Function to get the number of primaries generated

DUNEatLapp.get_rate_neutron_captures_Ar — Function

function getrateneutroncapturesAr(myfile::String,nameprimary::String,fidu::Real) function to get the rate of neutrons that are captured in LAr and in a fiducial volume of LAr It accepts a String for myfile, a String for nameprimary and a Real in centimeters for fidu. It returns two floats.

DUNEatLapp.greet_DUNEatLapp — Method

function greet_DUNEatLapp() Function to test if the package works There is no entry and the output is Hello DUNEatLapp!